Refactor PostList mess
